Windows project was updated. Some obsolete resources were removed.

+++ source/source/MinCOMLib/Concurrency/Core/IThread.win32.cpp
@@ -8,7 +8,7 @@ namespace MinCOM
 		return 0;
 	}
 
-    int IThread::GetSelfId()
+    longlong IThread::GetSelfId()
     {
         return ::GetCurrentThreadId();
     }
